Move vfs_hash_remove() calls earlier in virtfs_cleanup().
From jouyl@:
The sequence currently in the virtfs_vget_common() for new vnode is:
 - New VP
 - Add fid of the related virtfs_node of vnode to the list.
 - Insert VP to VFS hash table

This change moves VP removal from VFS hash table the earliest, since
once an np is associated with the vnode, VFS hash insertion is bound to
happen, modulo the case of failed insmntque().

This is for making sure the vnode is cleaned up properly even if a
vfs_hash_insert() call returns an existing vnode. Otherwise it stays on
the vfs hash table.

PR: 1695867
Change-Id: Ib19986ca6a96ed2095ee09a42110b937efe37e94