Remove unnecessary virtfs_cleanup in virtfs_close_session()
Since prior to the call into virtfs_cleanup, vflush() is already called,
the check and free here is simply unnecessary and in practise is never
triggered.

joyul@ spotted that the check and free is buggy after all as it calls
into virtfs_cleanup() which will lead to vses mtx being locked twice.

Change-Id: I773e7e46c743b7e5ff66ee3de7da04b973216297
PR:	1695867